fix: correct type annotations for labels and t_eval in PR #244
- Change _render_analysis_plots labels parameter from list[str] to
  tuple[str, ...] to match Palette.SEG_LABELS/BENCH_LABELS types
- Add dtype=np.float64 to np.linspace for t_eval to satisfy
  build_result_object's NDArray[np.float64] type constraint

+++ src/movement_optimizer/trajectory/optimizer.py
@@ -109,7 +109,7 @@ class TrajectoryOptimizer:
         if self.q_via is not None:
             n_ctrl += 1
         self.t_ctrl = np.linspace(0, self.duration, n_ctrl)
-        self.t_eval = np.linspace(0, self.duration, self.n_eval)
+        self.t_eval = np.linspace(0, self.duration, self.n_eval, dtype=np.float64)
 
     def build_splines(self, x: NDArray) -> list[CubicSpline]:
         """Build cubic splines from the flat optimisation vector *x*."""
